pyproject.toml
[project]
name = "placesapi-and-customsearchapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.161.0",
    "googlemaps>=4.10.0",
    "python-dotenv>=1.0.1",
]
